refactor(runtime): split vm_comparison_ops.rs into cohesive modules (§7-8)
`vm_comparison_ops.rs` was 1489 lines, over the 500-line module limit. Split it
into four files along its natural seams, each under 500 lines, with no behavior
change (the set of fn signatures is identical before/after, modulo visibility):

- `vm_comparison_ops.rs` (476) — the module-level numeric/value comparison free
  helpers (`cmp_values`, `range_cmp`, `is_nan_value`, range/list expansion, …)
  and the numeric comparison ops (`exec_num_eq`/`ne`/`lt`/`le`/`gt`/`ge`,
  `exec_approx_eq`).
- `vm_comparison_container_ops.rs` (233) — container-identity (`=:=`) ops
  (`exec_container_eq_op` and its named/indexed/raw variants + helpers).
- `vm_comparison_order_ops.rs` (494) — string comparison ops plus ordering
  (`spaceship_ordering`, `exec_spaceship`/`cmp`/`coll`/`unicmp`/`leg`/`eqv`,
  strict eq/ne, before/after).
- `vm_smartmatch_ops.rs` (303) — smartmatch (`~~`, `exec_smart_match_expr_op`)
  and divisibility (`%%`) ops.

The ordering module reads several comparison free helpers from
`vm_comparison_ops` (bumped `fn`→`pub(super)` and imported); `spaceship_ordering`
and `numeric_spaceship_ordering` are bumped `pub(super)` for the mutual
numeric↔ordering references. Eighth slice of the §7-8 file-splitting sweep.
