+++ src/vm.rs
@@ -136,7 +136,9 @@ mod vm_call_method_mut_ops;
 mod vm_call_method_ops;
 mod vm_closure_dispatch;
 mod vm_coerce_concat_ops;
+mod vm_comparison_container_ops;
 mod vm_comparison_ops;
+mod vm_comparison_order_ops;
 mod vm_control_ops;
 mod vm_data_ops;
 mod vm_dispatch_helpers;
@@ -159,6 +161,7 @@ mod vm_react_loop;
 mod vm_register_ops;
 mod vm_set_ops;
 pub(crate) mod vm_smart_match;
+mod vm_smartmatch_ops;
 pub(crate) mod vm_stats;
 pub(crate) mod vm_string_regex_ops;
 mod vm_value_helpers;
