+++ src/vm/vm_comparison_ops.rs
@@ -34,7 +34,7 @@ pub(super) fn value_to_f64(v: &Value) -> f64 {
     runtime::to_float_value(v).unwrap_or(0.0)
 }
 
-fn is_rationalish(v: &Value) -> bool {
+pub(super) fn is_rationalish(v: &Value) -> bool {
     matches!(
         v,
         Value::Rat(_, _) | Value::FatRat(_, _) | Value::BigRat(_, _)
@@ -43,7 +43,7 @@ fn is_rationalish(v: &Value) -> bool {
 
 /// Check if a value is NaN (for numeric comparison semantics).
 /// In Raku, Rat(0,0) and FatRat(0,0) behave like NaN.
-fn is_nan_value(v: &Value) -> bool {
+pub(super) fn is_nan_value(v: &Value) -> bool {
     match v {
         Value::Num(n) => n.is_nan(),
         Value::Rat(0, 0) | Value::FatRat(0, 0) => true,
@@ -54,7 +54,7 @@ fn is_nan_value(v: &Value) -> bool {
 
 /// Check if a value is Inf (positive infinity).
 /// Includes Rat(n,0)/FatRat(n,0) where n > 0.
-fn is_pos_inf(v: &Value) -> bool {
+pub(super) fn is_pos_inf(v: &Value) -> bool {
     match v {
         Value::Num(n) => n.is_infinite() && n.is_sign_positive(),
         Value::Rat(n, 0) | Value::FatRat(n, 0) => *n > 0,
@@ -65,7 +65,7 @@ fn is_pos_inf(v: &Value) -> bool {
 
 /// Check if a value is -Inf (negative infinity).
 /// Includes Rat(n,0)/FatRat(n,0) where n < 0.
-fn is_neg_inf(v: &Value) -> bool {
+pub(super) fn is_neg_inf(v: &Value) -> bool {
     match v {
         Value::Num(n) => n.is_infinite() && n.is_sign_negative(),
         Value::Rat(n, 0) | Value::FatRat(n, 0) => *n < 0,
@@ -98,7 +98,7 @@ fn extract_range_parts(v: &Value) -> Option<(Value, Value, bool, bool)> {
 
 /// Compare two ranges using Raku cmp semantics:
 /// min first, then excludes-min (False < True), then max, then excludes-max (True < False).
-fn range_cmp(left: &Value, right: &Value) -> std::cmp::Ordering {
+pub(super) fn range_cmp(left: &Value, right: &Value) -> std::cmp::Ordering {
     let (l_start, l_end, l_excl_start, l_excl_end) = extract_range_parts(left).unwrap();
     let (r_start, r_end, r_excl_start, r_excl_end) = extract_range_parts(right).unwrap();
 
@@ -123,7 +123,7 @@ fn range_cmp(left: &Value, right: &Value) -> std::cmp::Ordering {
 }
 
 /// Recursively apply cmp semantics to two values (for use in list element comparison).
-fn cmp_values(left: &Value, right: &Value) -> std::cmp::Ordering {
+pub(super) fn cmp_values(left: &Value, right: &Value) -> std::cmp::Ordering {
     // Handle NaN
     if is_nan_value(left) || is_nan_value(right) {
         return left.to_string_value().cmp(&right.to_string_value());
@@ -165,7 +165,7 @@ fn cmp_values(left: &Value, right: &Value) -> std::cmp::Ordering {
 }
 
 /// Get list elements from a value (Array, Seq, List, Slip, etc.)
-fn get_list_elements(v: &Value) -> Option<&[Value]> {
+pub(super) fn get_list_elements(v: &Value) -> Option<&[Value]> {
     match v {
         Value::Array(elems, _) => Some(elems.as_slice()),
         Value::Seq(elems) | Value::HyperSeq(elems) | Value::RaceSeq(elems) => {
@@ -177,12 +177,12 @@ fn get_list_elements(v: &Value) -> Option<&[Value]> {
 }
 
 /// Check if a value is any range variant.
-fn is_range_value(v: &Value) -> bool {
+pub(super) fn is_range_value(v: &Value) -> bool {
     extract_range_parts(v).is_some()
 }
 
 /// Expand a range value into a list of integer values (for list-based comparison).
-fn expand_range_to_list(v: &Value) -> Vec<Value> {
+pub(super) fn expand_range_to_list(v: &Value) -> Vec<Value> {
     match v {
         Value::Range(a, b) => (*a..=*b).map(Value::Int).collect(),
         Value::RangeExcl(a, b) => (*a..*b).map(Value::Int).collect(),
@@ -230,7 +230,7 @@ impl Interpreter {
         })
     }
 
-    fn numeric_spaceship_ordering(
+    pub(super) fn numeric_spaceship_ordering(
         left: &Value,
         right: &Value,
     ) -> Result<std::cmp::Ordering, RuntimeError> {
@@ -472,1018 +472,5 @@ impl Interpreter {
         let result = approx_f64(lr, rr) && approx_f64(li, ri);
         self.stack.push(Value::Bool(result));
         Ok(())
-    }
-
-    /// Container identity (`=:=`).
-    ///
-    /// `flags` encodes whether operands are provably fresh containers:
-    ///   bit 0 = left operand is a fresh container (e.g. array index),
-    ///   bit 1 = right operand is a fresh container.
-    ///
-    /// When at least one operand is a fresh container and both values
-    /// are non-reference types (no `Arc` identity), the two stack
-    /// values can never be the same container, so we return `False`.
-    /// Reference types (Array, Hash, Sub, Instance, …) have `Arc`
-    /// pointer identity which `values_identical` already checks.
-    pub(super) fn exec_container_eq_op(&mut self, flags: u8) {
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        let any_fresh = flags != 0;
-        let result = if any_fresh && Self::is_value_non_reference(&left, &right) {
-            // A fresh container (e.g. from `[$x][0]`) holding a non-reference
-            // value can never be the same container as any other expression.
-            false
-        } else {
-            crate::runtime::values_identical(&left, &right)
-        };
-        self.stack.push(Value::Bool(result));
-    }
-
-    /// Returns `true` when **both** values are simple, non-reference
-    /// types where stack copies can never carry container identity
-    /// (Int, Str, Bool, Nil, Rat, …).
-    ///
-    /// Note: `Package` is NOT included because type objects are singletons —
-    /// two variables holding the same Package should be considered `=:=`.
-    fn is_value_non_reference(left: &Value, right: &Value) -> bool {
-        fn is_non_ref(v: &Value) -> bool {
-            matches!(
-                v,
-                Value::Int(_)
-                    | Value::BigInt(_)
-                    | Value::Num(_)
-                    | Value::Str(_)
-                    | Value::Bool(_)
-                    | Value::Nil
-                    | Value::Rat(..)
-                    | Value::FatRat(..)
-                    | Value::BigRat(..)
-                    | Value::Complex(..)
-                    | Value::Whatever
-                    | Value::HyperWhatever
-                    | Value::Enum { .. }
-                    | Value::Version { .. }
-            )
-        }
-        is_non_ref(left) && is_non_ref(right)
-    }
-
-    /// Container identity (`=:=`) when both operands are named variables.
-    /// Resolves alias chains to find the binding root of each name, then:
-    /// - If both roots are the same name, they share a container → True.
-    /// - For reference types (Array, Hash, Sub, Instance), use Arc pointer identity.
-    /// - Otherwise → False (distinct scalar containers).
-    pub(super) fn exec_container_eq_named_op(
-        &mut self,
-        code: &CompiledCode,
-        left_name_idx: u32,
-        right_name_idx: u32,
-    ) {
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        let left_name = Self::const_str(code, left_name_idx);
-        let right_name = Self::const_str(code, right_name_idx);
-
-        // Resolve alias chains to find the binding root of each variable.
-        // =:= tests container identity — two named variables are identical
-        // only when they share the same binding root (via `:=`).
-        let left_root = self.resolve_alias_root(left_name);
-        let right_root = self.resolve_alias_root(right_name);
-
-        let result = if left_root == right_root {
-            // Same binding root → same container.
-            true
-        } else {
-            // Different binding roots. For most types, this means different
-            // containers → False. But singleton/unboxed types like Package
-            // (type objects) and Sub have no container semantics — two
-            // variables holding the same Package are considered =:=.
-            // Instance values (user objects) are NOT singletons: assignment
-            // copies the reference but creates a new container, so =:= is False.
-            match (&left, &right) {
-                (Value::Package(a), Value::Package(b)) => a == b,
-                (Value::Sub(a), Value::Sub(b)) => std::sync::Arc::ptr_eq(a, b),
-                (Value::WeakSub(a), Value::WeakSub(b)) => a.ptr_eq(b),
-                _ => false,
-            }
-        };
-        self.stack.push(Value::Bool(result));
-    }
-
-    /// Walk the `__mutsu_sigilless_alias::` chain to find the ultimate
-    /// binding root for a variable name.
-    fn resolve_alias_root(&self, name: &str) -> String {
-        let mut current = name.to_string();
-        let mut seen = std::collections::HashSet::new();
-        while seen.insert(current.clone()) {
-            let key = format!("__mutsu_sigilless_alias::{}", current);
-            if let Some(Value::Str(next)) = self.env().get(&key) {
-                current = next.to_string();
-            } else {
-                break;
-            }
-        }
-        current
-    }
-
-    /// Container identity (`=:=`) for indexed expressions (array/hash elements).
-    /// A `:=`-bound element holds a shared `ContainerRef` cell, so two bound
-    /// elements are the same container iff they hold the same cell `Arc`.
-    pub(super) fn exec_container_eq_indexed_op(
-        &mut self,
-        code: &CompiledCode,
-        left_name_idx: u32,
-        right_name_idx: u32,
-    ) {
-        let _right = self.stack.pop().unwrap();
-        let _left = self.stack.pop().unwrap();
-        let left_source = Self::const_str(code, left_name_idx);
-        let right_source = Self::const_str(code, right_name_idx);
-
-        // Same cell `Arc` at both slots → same container.
-        if let (Some(Value::ContainerRef(l)), Some(Value::ContainerRef(r))) = (
-            self.raw_element_at_encoded(left_source),
-            self.raw_element_at_encoded(right_source),
-        ) && Arc::ptr_eq(&l, &r)
-        {
-            self.stack.push(Value::Bool(true));
-            return;
-        }
-        // Same encoded source (e.g. @a[1] =:= @a[1])
-        let result = left_source == right_source;
-        self.stack.push(Value::Bool(result));
-    }
-
-    /// Container identity (`=:=`) using raw container values.
-    /// Compares HashEntryRef values to check if they
-    /// reference the same hash slot.
-    pub(super) fn exec_container_eq_raw_op(&mut self) {
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        let result =
-            Self::containers_same_slot(&left, &right) || Self::containers_same_slot(&right, &left);
-        self.stack.push(Value::Bool(result));
-    }
-
-    /// Check if two raw container values (HashEntryRef)
-    /// point to the same hash slot.
-    fn containers_same_slot(a: &Value, b: &Value) -> bool {
-        // Phase 2 element container: a `ContainerRef` cell has container identity
-        // by its `Arc<Mutex>`. Two cells are the same container only if the same
-        // Arc; a cell is never the same container as a non-cell value. (Must be
-        // checked here, not in `values_identical`, which `===` uses to compare
-        // *values* and therefore must read through a cell.)
-        match (a, b) {
-            (Value::ContainerRef(x), Value::ContainerRef(y)) => return Arc::ptr_eq(x, y),
-            (Value::ContainerRef(cell), other) | (other, Value::ContainerRef(cell)) => {
-                // A `:=`-bound hash element is promoted to a `ContainerRef` cell
-                // (Phase 2 Stage 1). The OTHER side may still be a HashEntryRef
-                // deferred token pointing at that *same* element (e.g. a
-                // deferred bind `my $b := %h<a><b>` whose element was later
-                // promoted when re-evaluated). They are the same container iff
-                // the element currently stored at that slot IS this cell.
-                if let Some((arc, key)) = Self::extract_hash_ref(other) {
-                    let ptr = Arc::as_ptr(&arc);
-                    if let Some(Value::ContainerRef(elem_cell)) =
-                        unsafe { (*ptr).get(key.as_str()) }
-                    {
-                        return Arc::ptr_eq(cell, elem_cell);
-                    }
-                }
-                return false;
-            }
-            _ => {}
-        }
-        // Extract (arc, key) from each side
-        let a_ref = Self::extract_hash_ref(a);
-        let b_ref = Self::extract_hash_ref(b);
-        if let (Some((a_arc, a_key)), Some((b_arc, b_key))) = (a_ref, b_ref) {
-            Arc::ptr_eq(&a_arc, &b_arc) && a_key == b_key
-        } else {
-            // Both are the same value identity (for non-hash-ref cases)
-            crate::runtime::values_identical(a, b)
-        }
-    }
-
-    /// Extract the `(terminal hash Arc, terminal key)` a `HashEntryRef` points
-    /// to, walking its `path` READ-ONLY. Returns `None` if any intermediate level
-    /// is missing or not a hash (the deferred path is not yet materialized, so it
-    /// has no stable container identity).
-    fn extract_hash_ref(val: &Value) -> Option<(Arc<crate::value::HashData>, String)> {
-        let Value::HashEntryRef { hash, path } = val else {
-            return None;
-        };
-        let mut cur = hash.clone();
-        for k in &path[..path.len() - 1] {
-            let ptr = Arc::as_ptr(&cur);
-            match unsafe { (*ptr).get(k.as_str()) } {
-                Some(Value::Hash(inner)) => cur = inner.clone(),
-                _ => return None,
-            }
-        }
-        Some((cur, path.last().unwrap().clone()))
-    }
-
-    /// For an encoded source like "@b\0idx\01" (or "%h\0idx\0k"), look up the
-    /// raw element value stored at that array/hash slot — without
-    /// decontainerizing, so a `:=`-bound `ContainerRef` cell is returned as-is
-    /// for identity comparison.
-    fn raw_element_at_encoded(&self, encoded: &str) -> Option<Value> {
-        let sep_pos = encoded.find("\x00idx\x00")?;
-        let var_name = &encoded[..sep_pos];
-        let idx_str = &encoded[sep_pos + 5..];
-        match self.env().get(var_name) {
-            Some(Value::Array(items, _)) => {
-                let i = idx_str.parse::<usize>().ok()?;
-                items.get(i).cloned()
-            }
-            Some(Value::Hash(hash)) => hash.get(idx_str).cloned(),
-            _ => None,
-        }
-    }
-
-    pub(super) fn exec_str_eq_op(&mut self) -> Result<(), RuntimeError> {
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        let result = self.eval_binary_with_junctions(left, right, |_, l, r| {
-            if Self::is_buf_value(&l) && Self::is_buf_value(&r) {
-                Ok(Value::Bool(
-                    Self::buf_cmp_bytes(&l, &r) == std::cmp::Ordering::Equal,
-                ))
-            } else {
-                Ok(Value::Bool(l.to_string_value() == r.to_string_value()))
-            }
-        })?;
-        self.stack.push(result);
-        Ok(())
-    }
-
-    pub(super) fn exec_str_ne_op(&mut self) -> Result<(), RuntimeError> {
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        // ne is a negation meta-operator shortcut for !eq.
-        // It first evaluates eq (which autothreads through junctions),
-        // then negates the boolean-collapsed result, always returning Bool.
-        let eq_result = self.eval_binary_with_junctions(left, right, |_, l, r| {
-            if Self::is_buf_value(&l) && Self::is_buf_value(&r) {
-                Ok(Value::Bool(
-                    Self::buf_cmp_bytes(&l, &r) == std::cmp::Ordering::Equal,
-                ))
-            } else {
-                Ok(Value::Bool(l.to_string_value() == r.to_string_value()))
-            }
-        })?;
-        self.stack.push(Value::Bool(!eq_result.truthy()));
-        Ok(())
-    }
-
-    fn buf_cmp_bytes(l: &Value, r: &Value) -> std::cmp::Ordering {
-        let lb = Self::extract_buf_bytes(l);
-        let rb = Self::extract_buf_bytes(r);
-        lb.cmp(&rb)
-    }
-
-    pub(super) fn exec_str_lt_op(&mut self) -> Result<(), RuntimeError> {
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        let result = self.eval_binary_with_junctions(left, right, |_, l, r| {
-            if Self::is_buf_value(&l) && Self::is_buf_value(&r) {
-                Ok(Value::Bool(
-                    Self::buf_cmp_bytes(&l, &r) == std::cmp::Ordering::Less,
-                ))
-            } else {
-                Ok(Value::Bool(l.to_string_value() < r.to_string_value()))
-            }
-        })?;
-        self.stack.push(result);
-        Ok(())
-    }
-
-    pub(super) fn exec_str_gt_op(&mut self) -> Result<(), RuntimeError> {
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        let result = self.eval_binary_with_junctions(left, right, |_, l, r| {
-            if Self::is_buf_value(&l) && Self::is_buf_value(&r) {
-                Ok(Value::Bool(
-                    Self::buf_cmp_bytes(&l, &r) == std::cmp::Ordering::Greater,
-                ))
-            } else {
-                Ok(Value::Bool(l.to_string_value() > r.to_string_value()))
-            }
-        })?;
-        self.stack.push(result);
-        Ok(())
-    }
-
-    pub(super) fn exec_str_le_op(&mut self) -> Result<(), RuntimeError> {
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        let result = self.eval_binary_with_junctions(left, right, |_, l, r| {
-            if Self::is_buf_value(&l) && Self::is_buf_value(&r) {
-                Ok(Value::Bool(
-                    Self::buf_cmp_bytes(&l, &r) != std::cmp::Ordering::Greater,
-                ))
-            } else {
-                Ok(Value::Bool(l.to_string_value() <= r.to_string_value()))
-            }
-        })?;
-        self.stack.push(result);
-        Ok(())
-    }
-
-    pub(super) fn exec_str_ge_op(&mut self) -> Result<(), RuntimeError> {
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        let result = self.eval_binary_with_junctions(left, right, |_, l, r| {
-            if Self::is_buf_value(&l) && Self::is_buf_value(&r) {
-                Ok(Value::Bool(
-                    Self::buf_cmp_bytes(&l, &r) != std::cmp::Ordering::Less,
-                ))
-            } else {
-                Ok(Value::Bool(l.to_string_value() >= r.to_string_value()))
-            }
-        })?;
-        self.stack.push(result);
-        Ok(())
-    }
-
-    fn spaceship_ordering(left: &Value, right: &Value) -> std::cmp::Ordering {
-        // Unwrap Mixin for comparison
-        let left = match left {
-            Value::Mixin(inner, _) => inner.as_ref(),
-            other => other,
-        };
-        let right = match right {
-            Value::Mixin(inner, _) => inner.as_ref(),
-            other => other,
-        };
-        match (left, right) {
-            (Value::Int(a), Value::Int(b)) => a.cmp(b),
-            (l, r)
-                if (is_rationalish(l) || is_rationalish(r))
-                    && matches!(
-                        l,
-                        Value::Int(_)
-                            | Value::BigInt(_)
-                            | Value::Rat(_, _)
-                            | Value::FatRat(_, _)
-                            | Value::BigRat(_, _)
-                    )
-                    && matches!(
-                        r,
-                        Value::Int(_)
-                            | Value::BigInt(_)
-                            | Value::Rat(_, _)
-                            | Value::FatRat(_, _)
-                            | Value::BigRat(_, _)
-                    ) =>
-            {
-                runtime::to_big_rat_parts(l)
-                    .zip(runtime::to_big_rat_parts(r))
-                    .and_then(|(a, b)| runtime::compare_big_rat_parts(a, b))
-                    .unwrap_or(std::cmp::Ordering::Equal)
-            }
-            (Value::Num(a), Value::Num(b)) => a.partial_cmp(b).unwrap_or(std::cmp::Ordering::Equal),
-            (Value::Int(a), Value::Num(b)) => (*a as f64)
-                .partial_cmp(b)
-                .unwrap_or(std::cmp::Ordering::Equal),
-            (Value::Num(a), Value::Int(b)) => a
-                .partial_cmp(&(*b as f64))
-                .unwrap_or(std::cmp::Ordering::Equal),
-            (
-                Value::Instance {
-                    attributes: left_attrs,
-                    ..
-                },
-                Value::Instance {
-                    attributes: right_attrs,
-                    ..
-                },
-            ) if left_attrs.contains_key("year")
-                && left_attrs.contains_key("month")
-                && left_attrs.contains_key("day")
-                && left_attrs.contains_key("hour")
-                && left_attrs.contains_key("minute")
-                && left_attrs.contains_key("second")
-                && left_attrs.contains_key("timezone")
-                && right_attrs.contains_key("year")
-                && right_attrs.contains_key("month")
-                && right_attrs.contains_key("day")
-                && right_attrs.contains_key("hour")
-                && right_attrs.contains_key("minute")
-                && right_attrs.contains_key("second")
-                && right_attrs.contains_key("timezone") =>
-            {
-                let (ly, lm, ld, lh, lmin, ls, ltz) =
-                    crate::builtins::methods_0arg::temporal::datetime_attrs(&(left_attrs).as_map());
-                let (ry, rm, rd, rh, rmin, rs, rtz) =
-                    crate::builtins::methods_0arg::temporal::datetime_attrs(
-                        &(right_attrs).as_map(),
-                    );
-                let left_instant =
-                    crate::builtins::methods_0arg::temporal::datetime_to_instant_leap_aware(
-                        ly, lm, ld, lh, lmin, ls, ltz,
-                    );
-                let right_instant =
-                    crate::builtins::methods_0arg::temporal::datetime_to_instant_leap_aware(
-                        ry, rm, rd, rh, rmin, rs, rtz,
-                    );
-                left_instant
-                    .partial_cmp(&right_instant)
-                    .unwrap_or(std::cmp::Ordering::Equal)
-            }
-            (l, r)
-                if matches!(
-                    l,
-                    Value::Int(_)
-                        | Value::BigInt(_)
-                        | Value::Num(_)
-                        | Value::Rat(_, _)
-                        | Value::FatRat(_, _)
-                        | Value::BigRat(_, _)
-                ) && matches!(
-                    r,
-                    Value::Int(_)
-                        | Value::BigInt(_)
-                        | Value::Num(_)
-                        | Value::Rat(_, _)
-                        | Value::FatRat(_, _)
-                        | Value::BigRat(_, _)
-                ) =>
-            {
-                value_to_f64(l)
-                    .partial_cmp(&value_to_f64(r))
-                    .unwrap_or(std::cmp::Ordering::Equal)
-            }
-            // Complex cmp: compare real parts first, then imaginary parts
-            // NaN sorts as Greater (More) in Raku
-            (Value::Complex(ar, ai), Value::Complex(br, bi)) => {
-                if ar.is_nan() || ai.is_nan() {
-                    return std::cmp::Ordering::Greater;
-                }
-                match ar.partial_cmp(br).unwrap_or(std::cmp::Ordering::Greater) {
-                    std::cmp::Ordering::Equal => {
-                        ai.partial_cmp(bi).unwrap_or(std::cmp::Ordering::Greater)
-                    }
-                    ord => ord,
-                }
-            }
-            // Complex vs Real: treat Real as Complex with im=0
-            (Value::Complex(ar, ai), _) => {
-                if ar.is_nan() || ai.is_nan() {
-                    return std::cmp::Ordering::Greater;
-                }
-                let br = value_to_f64(right);
-                match ar.partial_cmp(&br).unwrap_or(std::cmp::Ordering::Greater) {
-                    std::cmp::Ordering::Equal => {
-                        ai.partial_cmp(&0.0).unwrap_or(std::cmp::Ordering::Greater)
-                    }
-                    ord => ord,
-                }
-            }
-            (_, Value::Complex(br, bi)) => {
-                if br.is_nan() || bi.is_nan() {
-                    return std::cmp::Ordering::Greater;
-                }
-                let ar = value_to_f64(left);
-                match ar.partial_cmp(br).unwrap_or(std::cmp::Ordering::Greater) {
-                    std::cmp::Ordering::Equal => 0.0f64
-                        .partial_cmp(bi)
-                        .unwrap_or(std::cmp::Ordering::Greater),
-                    ord => ord,
-                }
-            }
-            (Value::Version { parts: ap, .. }, Value::Version { parts: bp, .. }) => {
-                runtime::version_cmp_parts(ap, bp)
-            }
-            // Enum values: compare by their integer value
-            (Value::Enum { value: av, .. }, Value::Enum { value: bv, .. }) => {
-                av.as_i64().cmp(&bv.as_i64())
-            }
-            // Range cmp Range (both must be ranges; Range vs List handled below)
-            (l, r) if is_range_value(l) && is_range_value(r) => range_cmp(l, r),
-            // Real cmp Range (where Real is not a list/range)
-            (_, r)
-                if is_range_value(r)
-                    && !is_range_value(left)
-                    && get_list_elements(left).is_none() =>
-            {
-                let as_range = Value::GenericRange {
-                    start: Arc::new(left.clone()),
-                    end: Arc::new(left.clone()),
-                    excl_start: false,
-                    excl_end: false,
-                };
-                range_cmp(&as_range, r)
-            }
-            // Range cmp Real (where Real is not a list/range)
-            (l, _)
-                if is_range_value(l)
-                    && !is_range_value(right)
-                    && get_list_elements(right).is_none() =>
-            {
-                let as_range = Value::GenericRange {
-                    start: Arc::new(right.clone()),
-                    end: Arc::new(right.clone()),
-                    excl_start: false,
-                    excl_end: false,
-                };
-                range_cmp(l, &as_range)
-            }
-            // List/Array cmp (including Range vs List: expand range to list)
-            (l, r)
-                if get_list_elements(l).is_some()
-                    || get_list_elements(r).is_some()
-                    || is_range_value(l)
-                    || is_range_value(r) =>
-            {
-                let l_vec;
-                let l_elems: &[Value] = if let Some(elems) = get_list_elements(l) {
-                    elems
-                } else if is_range_value(l) {
-                    l_vec = expand_range_to_list(l);
-                    &l_vec
-                } else {
-                    l_vec = vec![l.clone()];
-                    &l_vec
-                };
-                let r_vec;
-                let r_elems: &[Value] = if let Some(elems) = get_list_elements(r) {
-                    elems
-                } else if is_range_value(r) {
-                    r_vec = expand_range_to_list(r);
-                    &r_vec
-                } else {
-                    r_vec = vec![r.clone()];
-                    &r_vec
-                };
-                for (le, re) in l_elems.iter().zip(r_elems.iter()) {
-                    let ord = cmp_values(le, re);
-                    if ord != std::cmp::Ordering::Equal {
-                        return ord;
-                    }
-                }
-                l_elems.len().cmp(&r_elems.len())
-            }
-            _ => left.to_string_value().cmp(&right.to_string_value()),
-        }
-    }
-
-    pub(super) fn exec_spaceship_op(&mut self) -> Result<(), RuntimeError> {
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        let result = self.eval_binary_with_junctions(left, right, |vm, l, r| {
-            let (l, r) = vm.coerce_numeric_bridge_pair(l, r)?;
-            // NaN <=> anything produces Nil (unordered)
-            if is_nan_value(&l) || is_nan_value(&r) {
-                return Ok(Value::Nil);
-            }
-            // Complex <=> Real: check $*TOLERANCE for imaginary part
-            let l = vm.coerce_complex_to_real_if_tolerant(&l)?;
-            let r = vm.coerce_complex_to_real_if_tolerant(&r)?;
-            let ord = Self::numeric_spaceship_ordering(&l, &r)?;
-            Ok(runtime::make_order(ord))
-        })?;
-        self.stack.push(result);
-        Ok(())
-    }
-
-    /// If a Complex value has an imaginary part within `$*TOLERANCE` (relative),
-    /// coerce it to its real part. Otherwise throw if it's Complex with a
-    /// significant imaginary component.
-    fn coerce_complex_to_real_if_tolerant(&mut self, val: &Value) -> Result<Value, RuntimeError> {
-        if let Value::Complex(re, im) = val {
-            if *im == 0.0 {
-                return Ok(Value::Num(*re));
-            }
-            let tolerance = loan_env!(self, get_dynamic_var("$*TOLERANCE"))
-                .ok()
-                .and_then(|v| runtime::to_float_value(&v))
-                .unwrap_or(1e-15);
-            let re_abs = re.abs();
-            if re_abs != 0.0 && im.abs() / re_abs <= tolerance {
-                Ok(Value::Num(*re))
-            } else {
-                Err(RuntimeError::new(
-                    "Cannot use <=> on Complex number with non-negligible imaginary part",
-                ))
-            }
-        } else {
-            Ok(val.clone())
-        }
-    }
-
-    pub(super) fn exec_before_after_op(&mut self, is_before: bool) {
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        let (left, right) = self
-            .coerce_numeric_bridge_pair(left.clone(), right.clone())
-            .unwrap_or((left, right));
-        let ord = Self::spaceship_ordering(&left, &right);
-        let result = if is_before {
-            ord == std::cmp::Ordering::Less
-        } else {
-            ord == std::cmp::Ordering::Greater
-        };
-        self.stack.push(Value::Bool(result));
-    }
-
-    pub(super) fn exec_cmp_op(&mut self) {
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        if Self::is_buf_value(&left) && Self::is_buf_value(&right) {
-            let ord = Self::buf_cmp_bytes(&left, &right);
-            self.stack.push(runtime::make_order(ord));
-            return;
-        }
-        // NaN in cmp context: compare as string "NaN"
-        if is_nan_value(&left) || is_nan_value(&right) {
-            let ord = left.to_string_value().cmp(&right.to_string_value());
-            self.stack.push(runtime::make_order(ord));
-            return;
-        }
-        // For lists, ranges, and mixed-type comparisons (e.g. Pair cmp Inf),
-        // use cmp_values which handles these cases correctly.
-        if get_list_elements(&left).is_some()
-            || get_list_elements(&right).is_some()
-            || is_range_value(&left)
-            || is_range_value(&right)
-            || is_pos_inf(&left)
-            || is_neg_inf(&left)
-            || is_pos_inf(&right)
-            || is_neg_inf(&right)
-        {
-            let ord = cmp_values(&left, &right);
-            self.stack.push(runtime::make_order(ord));
-            return;
-        }
-        let (left, right) = self
-            .coerce_numeric_bridge_pair(left.clone(), right.clone())
-            .unwrap_or((left, right));
-        let ord = Self::spaceship_ordering(&left, &right);
-        self.stack.push(runtime::make_order(ord));
-    }
-
-    pub(super) fn exec_coll_op(&mut self) {
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        let left_s = left.to_string_value();
-        let right_s = right.to_string_value();
-
-        // Get $*COLLATION settings
-        let settings = self.get_collation_settings();
-        let result = crate::builtins::collation::coll_compare(&left_s, &right_s, &settings);
-        self.stack.push(result);
-    }
-
-    pub(super) fn exec_unicmp_op(&mut self) {
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        let left_s = left.to_string_value();
-        let right_s = right.to_string_value();
-
-        // Unlike `coll`, `unicmp` always uses the default Unicode collation and
-        // is not influenced by the `$*COLLATION` dynamic variable.
-        let settings = crate::builtins::collation::CollationSettings::default();
-        let result = crate::builtins::collation::coll_compare(&left_s, &right_s, &settings);
-        self.stack.push(result);
-    }
-
-    pub(super) fn exec_leg_op(&mut self) {
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        let ord = left.to_string_value().cmp(&right.to_string_value());
-        self.stack.push(runtime::make_order(ord));
-    }
-
-    pub(super) fn exec_strict_eq_op(&mut self) -> Result<(), RuntimeError> {
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        let result = self.eval_binary_with_junctions(left, right, |_, l, r| {
-            Ok(Value::Bool(runtime::values_identical(&l, &r)))
-        })?;
-        self.stack.push(result);
-        Ok(())
-    }
-
-    pub(super) fn exec_strict_ne_op(&mut self) -> Result<(), RuntimeError> {
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        // !== is a negation meta-operator applied to ===.
-        // It first evaluates === (which autothreads through junctions),
-        // then negates the boolean-collapsed result, always returning Bool.
-        let eq_result = self.eval_binary_with_junctions(left, right, |_, l, r| {
-            Ok(Value::Bool(runtime::values_identical(&l, &r)))
-        })?;
-        self.stack.push(Value::Bool(!eq_result.truthy()));
-        Ok(())
-    }
-
-    pub(super) fn exec_eqv_op(&mut self) -> Result<(), RuntimeError> {
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        let result =
-            self.eval_binary_with_junctions(left, right, |_, l, r| Ok(Value::Bool(l.eqv(&r))))?;
-        self.stack.push(result);
-        Ok(())
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    pub(super) fn exec_smart_match_expr_op(
-        &mut self,
-        code: &CompiledCode,
-        ip: &mut usize,
-        rhs_end: u32,
-        negate: bool,
-        lhs_var: &Option<String>,
-        rhs_is_match_regex: bool,
-        lhs_is_literal: bool,
-        rhs_pure_regex: bool,
-        compiled_fns: &HashMap<String, CompiledFunction>,
-    ) -> Result<(), RuntimeError> {
-        let left = self.stack.pop().unwrap();
-        let rhs_start = *ip + 1;
-        let rhs_end = rhs_end as usize;
-        // Slice 6.3 step 2: clear the regex engine's caller-variable write log so
-        // that, after the match, a non-empty `pending_local_updates` reliably means
-        // *this* match's embedded `{ }` code blocks wrote a caller variable by name
-        // (the engine records them there but nothing consumes the log otherwise).
-        self.pending_local_updates.clear();
-        let saved_topic = self.env().get("_").cloned();
-        self.env_mut().insert("_".to_string(), left.clone());
-        // While the RHS runs, `$_` is *aliased* to the LHS variable (`$x ~~ s///`
-        // topicalizes `$x`). A destructive `s///`/`tr///` checks `$_`'s readonly
-        // status, but the enclosing scope may have marked `_` readonly for an
-        // unrelated reason — e.g. a `for ^N { $x ~~ s/// }` loop marks its *topic*
-        // `_` readonly. The match's `_` must instead reflect the *aliased
-        // variable*'s mutability: mutable `my $x` allows the write, a readonly
-        // param still blocks it. Save and override `_`'s readonly flag for the
-        // duration when the LHS is an explicit variable other than the topic.
-        let topic_ro_override = match lhs_var {
-            Some(v) if v != "_" => {
-                let saved = self.is_readonly("_");
-                let bare = v.trim_start_matches(['$', '@', '%', '&']);
-                let target_ro = self.is_readonly(v) || self.is_readonly(bare);
-                if target_ro {
-                    self.mark_readonly("_");
-                } else {
-                    self.unmark_readonly("_");
-                }
-                Some(saved)
-            }
-            _ => None,
-        };
-        // Clear the env-dirty flag before overwriting env with local values for
-        // regex interpolation. The reverse env->locals pull this used to perform
-        // was removed in Stage 3 (the interpreter-bridge writes it covered, e.g.
-        // EVAL modifying a `$GLOBAL::` variable, are now reconciled by precise
-        // write-throughs at their own sites).
-        self.sync_regex_interpolation_env_from_locals(code);
-        let saved_in_smartmatch_rhs = self.in_smartmatch_rhs;
-        self.in_smartmatch_rhs = true;
-        self.transliterate_in_smartmatch = false;
-        self.substitution_in_smartmatch = false;
-        let rhs_run = self.run_range(code, rhs_start, rhs_end, compiled_fns);
-        self.in_smartmatch_rhs = saved_in_smartmatch_rhs;
-        let was_transliterate = self.transliterate_in_smartmatch;
-        let was_substitution = self.substitution_in_smartmatch;
-        self.transliterate_in_smartmatch = false;
-        self.substitution_in_smartmatch = false;
-        // Restore the topic's readonly flag (overridden above for an aliased LHS
-        // variable) on every path, including the error path below.
-        if let Some(saved) = topic_ro_override {
-            if saved {
-                self.mark_readonly("_");
-            } else {
-                self.unmark_readonly("_");
-            }
-        }
-        rhs_run?;
-        let right = self.stack.pop().unwrap_or(Value::Nil);
-        // A destructive `s///`/`tr///` that actually matched against a string
-        // literal has no writable container to update, so Raku throws
-        // X::Assignment::RO (e.g. `'abc' ~~ s/b/g/`). A non-matching attempt is
-        // a no-op and does not throw.
-        if lhs_is_literal && (was_substitution || was_transliterate) && right.truthy() {
-            // Restore the topic before propagating the error.
-            if let Some(v) = saved_topic {
-                self.env_mut().insert("_".to_string(), v);
-            } else {
-                self.env_mut().remove("_");
-            }
-            return Err(RuntimeError::assignment_ro(Some("Str")));
-        }
-        // When the LHS alias *is* the topic itself (`$_ ~~ s///`), the
-        // substitution-modified topic must persist — restoring `saved_topic`
-        // below would clobber it. Skip the restore in that case.
-        let lhs_is_topic = lhs_var.as_deref() == Some("_");
-        if let Some(var_name) = lhs_var {
-            let modified_topic = self.env().get("_").cloned().unwrap_or(Value::Nil);
-            self.env_mut()
-                .insert(var_name.clone(), modified_topic.clone());
-            // Reverse write-through: if the lhs alias names a compiled local slot,
-            // mirror the (possibly substitution-modified) topic into it so the
-            // caller sees it without an O(locals) sync_locals_from_env pull.
-            self.update_local_if_exists(code, var_name, &modified_topic);
-            // The env insert above bypasses `set_env_with_main_alias`, and
-            // `update_local_if_exists` only touches the *current* frame's slot —
-            // when `$x ~~ s///` runs inside an EVAL/carrier, `var_name` is an outer
-            // lexical with no current-frame slot, so flag it for the reverse sync
-            // (carrier log + env_dirty). Otherwise a carrier dropping its blanket
-            // net loses the `~~ s///` writeback (Slice C', open-question #2).
-            self.note_caller_env_write(var_name);
-            // When the topic itself was modified (`$_ ~~ s///` inside a
-            // `given $x` / `with $x` block), `$_` aliases the source scalar
-            // variable, so the in-place substitution must propagate back to it —
-            // mirroring the whole-topic writeback the `$_ = ...` assign path does.
-            if lhs_is_topic
-                && let Some(ref source_var) = self.topic_source_var
-                && !source_var.starts_with('@')
-                && !source_var.starts_with('%')
-            {
-                let sv = source_var.clone();
-                self.set_env_with_main_alias(&sv, modified_topic.clone());
-                self.update_local_if_exists(code, &sv, &modified_topic);
-            }
-        }
-        if !lhs_is_topic {
-            if let Some(v) = saved_topic {
-                self.env_mut().insert("_".to_string(), v);
-            } else {
-                self.env_mut().remove("_");
-            }
-        }
-        // When RHS was a transliterate (tr///), return the result directly.
-        // In Raku, $x ~~ tr/a/b/ returns a StrDistance that stringifies to the after-string.
-        let out = if was_transliterate {
-            if negate {
-                // !~~ tr/// — negate the truthiness of the result
-                Value::Bool(!right.truthy())
-            } else {
-                right
-            }
-        } else if was_substitution {
-            // s/// returns Match on success, False on failure.
-            // For !~~, negate the boolean result.
-            if negate {
-                Value::Bool(!right.truthy())
-            } else {
-                right
-            }
-        } else if negate {
-            // Smartmatch must NOT force lazy values (lazy ~~ anything → False)
-            self.eval_smartmatch_with_junctions_ex(left, right, true, rhs_is_match_regex)?
-        } else {
-            self.eval_smartmatch_with_junctions_ex(left, right, false, rhs_is_match_regex)?
-        };
-        self.stack.push(out);
-        // Slice 6.3 step 2 — precise env_dirty for smartmatch. Skip the
-        // O(caller-locals) re-sync only when this was a side-effect-free match:
-        //   * a plain `Value::Regex` literal (rhs_pure_regex) — excludes
-        //     RegexWithAdverbs (`:pos`/`:g` carry state), named/Sub regexes, and
-        //     value smartmatch, all of which can write caller state by name;
-        //   * not a substitution / transliteration (those mutate the topic alias);
-        //   * the match ran no embedded `{ }` code block that wrote a caller
-        //     variable (the regex engine logs those in `pending_local_updates`);
-        //   * `$/` is not itself a compiled local slot here (`my $/` — e.g.
-        //     continued matching reads it back across calls).
-        // A plain regex match otherwise only writes `$/`/captures, which are
-        // special vars read by name, never a caller local slot, so no pull.
-        let wrote_caller_via_code = !self.pending_local_updates.is_empty();
-        let match_var_is_local = code.locals.iter().any(|n| n == "/");
-        let pure = rhs_pure_regex
-            && !was_substitution
-            && !was_transliterate
-            && !wrote_caller_via_code
-            && !match_var_is_local;
-        // Slice F: a match writes caller lexicals straight into `env` — the match
-        // variable `$/`, numbered captures (`$0`/`$1`/…), and an embedded `{ }` /
-        // `:my` / `:let` block's writes (logged in `pending_local_updates`). Write
-        // those through to the caller's local slots here so the slot stays
-        // coherent without the reverse `sync_locals_from_env` pull. `env_dirty`
-        // below is kept as a net (reverse-sync ON behaves identically); with it
-        // OFF this write-through is what keeps the value correct.
-        if !pure {
-            let written: std::collections::HashSet<String> = self
-                .pending_local_updates
-                .iter()
-                .map(|(n, _)| n.clone())
-                .collect();
-            self.writeback_match_locals(code, &written);
-            // env_dirty substrate (docs/captured-outer-cell-sharing.md §10): an
-            // embedded `{ }` / `:my` / `:let` block writes a caller lexical that is
-            // NOT a slot of *this* frame — e.g. a regex run inside `sub do-match { … }`
-            // mutates a lexical captured from the sub's caller. `writeback_match_locals`
-            // only reaches this frame's slots, and the owning slot lives one or more
-            // frames up, so record those names for the retain-on-miss caller-var
-            // writeback; the call site that returns to the owning frame drains them
-            // (`apply_pending_caller_var_writeback`). This is what keeps the slot
-            // coherent once the blanket/precise reconcile is gone (double-OFF). Only
-            // armed under boxing; the default build's blanket reconcile covers it.
-            for name in &written {
-                let is_match_name =
-                    name == "/" || (!name.is_empty() && name.bytes().all(|b| b.is_ascii_digit()));
-                if is_match_name || self.find_local_slot(code, name).is_some() {
-                    continue;
-                }
-                self.record_caller_var_writeback(name);
-            }
-        }
-        self.pending_local_updates.clear();
-        // env_dirty substrate (docs/captured-outer-cell-sharing.md §10): a custom
-        // HOW `type_check`/`accepts_type`/`find_method` invoked by this smartmatch
-        // (`$obj ~~ $custom-type`) may mutate a captured-outer caller scalar (e.g.
-        // `++$counter`); the slow HOW dispatch records those names into the
-        // retain-on-miss caller-var writeback but never sets `env_dirty`. Drain it
-        // here so the owning slot is refreshed before the next statement reads it
-        // (critical for the read-modify-write `++$counter`). No-op in default
-        // builds (the recording is gated on cell_boxing_active); cross-frame
-        // names recorded by the regex path above are retained for the caller.
-        self.apply_pending_caller_var_writeback(code);
-        *ip = rhs_end;
-        Ok(())
-    }
-
-    pub(super) fn exec_scalarize_regex_match_result_op(&mut self) -> Result<(), RuntimeError> {
-        let value = self.stack.pop().unwrap_or(Value::Nil);
-        let scalarized = match value {
-            Value::Nil => Value::Int(0),
-            Value::Array(items, _) => Value::Int(items.len() as i64),
-            Value::Seq(items)
-            | Value::HyperSeq(items)
-            | Value::RaceSeq(items)
-            | Value::Slip(items) => Value::Int(items.len() as i64),
-            Value::Capture { positional, .. } => Value::Int(positional.len() as i64),
-            Value::Instance {
-                class_name,
-                attributes,
-                ..
-            } if class_name == "Match" => {
-                if let Some(Value::Array(items, _)) = attributes.as_map().get("list") {
-                    Value::Int(items.len() as i64)
-                } else {
-                    Value::Int(1)
-                }
-            }
-            other if other.truthy() => Value::Int(1),
-            _ => Value::Int(0),
-        };
-        self.stack.push(scalarized);
-        Ok(())
-    }
-
-    pub(super) fn exec_divisible_by_op(&mut self) -> Result<(), RuntimeError> {
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        // Thread over junctions
-        if matches!(left, Value::Junction { .. }) || matches!(right, Value::Junction { .. }) {
-            let result = self
-                .eval_binary_with_junctions(left, right, |vm, l, r| vm.divisible_by_values(l, r))?;
-            self.stack.push(result);
-            return Ok(());
-        }
-        let result = self.divisible_by_values(left.clone(), right)?;
-        self.stack.push(result);
-        Ok(())
-    }
-
-    fn divisible_by_values(&self, left: Value, right: Value) -> Result<Value, RuntimeError> {
-        Ok(Value::Bool(self.is_divisible(left, right)?))
-    }
-
-    /// `$a %% $b` is `$a % $b == 0`. Compute the modulo with the exact-rational
-    /// semantics of `arith_mod` (so Rat/Num/BigInt operands work, not just Int),
-    /// then test the remainder for zero. A zero divisor reports the dividend and
-    /// `infix:<%%>`, matching Rakudo.
-    fn is_divisible(&self, left: Value, right: Value) -> Result<bool, RuntimeError> {
-        let (l, r) = runtime::coerce_numeric(left.clone(), right);
-        if !r.truthy() {
-            return Err(RuntimeError::numeric_divide_by_zero_full(
-                Some(left),
-                Some("infix:<%%>"),
-            ));
-        }
-        let remainder = crate::builtins::arith_mod(l, r)?;
-        Ok(!remainder.truthy())
-    }
-
-    pub(super) fn exec_not_divisible_by_op(&mut self) -> Result<(), RuntimeError> {
-        let right = self.stack.pop().unwrap();
-        let left = self.stack.pop().unwrap();
-        // Thread over junctions
-        if matches!(left, Value::Junction { .. }) || matches!(right, Value::Junction { .. }) {
-            let result = self.eval_binary_with_junctions(left, right, |vm, l, r| {
-                vm.not_divisible_by_values(l, r)
-            })?;
-            self.stack.push(result);
-            return Ok(());
-        }
-        let result = self.not_divisible_by_values(left.clone(), right)?;
-        self.stack.push(result);
-        Ok(())
-    }
-
-    fn not_divisible_by_values(&self, left: Value, right: Value) -> Result<Value, RuntimeError> {
-        Ok(Value::Bool(!self.is_divisible(left, right)?))
     }
 }
